requirements.txt
albumentations>0.5.0
Click>=7.0
defusedxml
flask
glymur
imagecodecs
jupyterlab
matplotlib
numpy
opencv-python==4.5.4.60
openslide-python==1.1.2
pandas
pillow
pyyaml>=5.1
requests
scikit-image==0.18.3
scikit-learn>=0.23.2
shapely
sphinx==4.1.2
tifffile
torch==1.9.1
torchvision==0.10.1
tqdm==4.60.0
umap-learn
zarr
